requirements.txt
streamlit
gtts
SpeechRecognition
pydub
audio-recorder-streamlit
chromadb
openpyxl
streamlit
google-cloud-speech
google-cloud-texttospeech
sounddevice
soundfile
numpy
streamlit
google-cloud-speech
google-cloud-texttospeech
streamlit-audiorecorder
pandas
chromadb
requests
openpyxl
st_audiorec